pipeline-manager: restructure API module
Restructures the API module such that endpoints are in a separate module
and the API server main function is split off from `mod.rs` into
`main.rs`. This commit only moves files and splits one, and makes edits
in accordance with these changes. This is in order to keep its scope
limited. It is the first in a line of self-contained PRs to improve the
API.
