+++ crates/pipeline-manager/src/api/endpoints/mod.rs
@@ -1,0 +1,5 @@
+pub mod api_key;
+pub mod config;
+pub mod http_io;
+pub mod metrics;
+pub mod pipeline;
